Created new relational entities and the dp/model objects to utilise them

+++ app/src/main/java/com/example/lexicon_memoria/database/entity/LexiconEntity.kt
@@ -23,7 +23,7 @@ import androidx.room.PrimaryKey
 data class LexiconEntity(
         @ColumnInfo(name="user_id") val userId: Long,
         val name: String,
-        @ColumnInfo(name = "creation_date") val creationDate: Int
+        @ColumnInfo(name = "creation_date") val creationDate: Long
 ) : BaseEntity() {
     @PrimaryKey(autoGenerate = true) override var id: Long = 0
